edge-daemon/requirements.txt
# dexterslab edge-daemon requirements
pyaudio==0.2.14
websockets==12.0
python-dotenv==1.0.1
